Rename backend, minor tweaks. Still not compiling
