+++ src/main.rs
@@ -1,7 +1,7 @@
 #![warn(clippy::all, rust_2018_idioms)]
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")] // hide console window on Windows in release
 
-mod backend;
+mod datamanager;
 mod category;
 mod colors;
 mod csvadapter;
